+++ R/context.R
@@ -102,7 +102,7 @@ cur_group_label <- function() {
   data <- mask$full_data()
   if(is_grouped_df(data) && nrow(data) > 0) {
     glue("group {id}: {label}", id = cur_group_id(), label = group_labels_details(cur_group()))
-  } else if (inherits(data, "rowwise_df") && nrow(data) > 0) {
+  } else if (is_rowwise_df(data) && nrow(data) > 0) {
     paste0("row ", cur_group_id())
   } else {
     ""