+++ NEWS.md
@@ -14,6 +14,7 @@
    and `slice_sample()`) now accept negative values for `n` and `prop` (#5961).
 
 * `slice()` now indicates which group produces an error (#5931).
+* New function `is_rowwise_df()`
 
 # dplyr 1.0.7